:recycle: Refactor code.
