--- src/components/client/AnimatePresence/AnimatePresence.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import { AnimatePresence as FMAP } from "framer-motion";
-
-const AnimatePresence = FMAP;
-
-export default AnimatePresence;
