--- src/components/client/AnimatePresence/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AnimatePresence";
