+++ src/lib/framer-motion/index.ts
@@ -1,0 +1,3 @@
+"use client";
+
+export * from "framer-motion";
