+++ src/lib/geist-ui/index.ts
@@ -1,0 +1,3 @@
+"use client";
+
+export * from "@geist-ui/core";
