+++ src/lib/react-redux/index.ts
@@ -1,0 +1,3 @@
+"use client";
+
+export * from "react-redux";
